ändrade namn på en metod

+++ src/sim/SimulatorStateADT.java
@@ -5,11 +5,11 @@ import java.util.Observable;
 public abstract class SimulatorStateADT extends Observable {
 	
 	protected boolean started = false;
-	protected int calculatedTime = 0;
+	protected double calculatedTime = 0;
 	
 	abstract protected void start();
 	
 	abstract protected void stop();
 	
-	protected abstract int calculateTime();
+	protected abstract double calculateTime();
 }
